backend/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.0.5
django-colorfield==0.4.3
django-cors-headers==3.8.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.1.1
drf-yasg==1.20.0
flake8==4.0.1
gunicorn==20.0.4
idna==3.2
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
pluggy==0.13.1
postgres==4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
psycopg2-pool==1.1
py==1.10.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
python-dotenv==0.19.2